Modified to make the training use the objective function of BCEL with Neg sample generation. The prediction is a score for each tuple , the labels are 1 for +ve tuples, and 0 for -ve tuples. For testing also, Neg samples are generated and rank based metrics calculated for +ve tuples among their corresponding -ve sample set.  Changes to be committed:i 	modified:   Code/HEHR_GNN/data_input.py 	modified:   Code/HEHR_GNN/evaluation.py 	modified:   Code/HEHR_GNN/graph_decoder.py 	modified:   Code/HEHR_GNN/graph_encoder.py 	modified:   Code/HEHR_GNN/main.py 	new file:   Code/HEHR_GNN/results/training_loss_dm_training_e1_r_toe2_test_e1_r_to_e2.txt 	new file:   Code/HEHR_GNN/results/training_loss_dm_training_e1_r_toe2_test_e2_r_to_e1.txt 	modified:   Code/HEHR_GNN/training.py 	new file:   Code/HEHR_GNN/training_data_prep.py

+++ data_input.py
@@ -132,7 +132,7 @@ class InputDataManager():
                 pr_tuple_id, q_pairs_id = self.tuple2ids(i,line)
                 stmt = np.concatenate((pr_tuple_id,q_pairs_id),axis=0)
                 statements.append(stmt)
-            return np.array(statements)
+            return np.array(statements,dtype=np.int64)
     
     """ def read_test(self, file_path):
         if not os.path.exists(file_path):
@@ -184,5 +184,7 @@ class InputDataManager():
         if not rel in self.rel2id:
             self.rel2id[rel] = len(self.rel2id)
         return self.rel2id[rel]
+    
+
